requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
httpx>=0.27.0
anthropic>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.0.0
